pyproject.toml
[tool.poetry]
name = "eyeliner"
version = "0.1.0"
description = "Automatic Longitudinal Image Registration using Fundus Landmarks"
authors = ["Aveturi13 <advaithveturi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.2"
torchvision = "^0.17.2"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.4"
kornia = "^0.7.2"
tqdm = "^4.66.2"
pandas = "^2.2.2"
scipy = "^1.13.0"
scikit-image = "^0.23.2"
monai = "^1.3.0"
patchify = "^0.2.3"
imageio = {extras = ["ffmpeg"], version = "^2.34.1"}
jupyter = "^1.0.0"
openpyxl = "^3.1.4"
moviepy = "^1.0.3"
python-pptx = "^0.6.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
